+++ chatbot/src/config.py
@@ -1,0 +1,116 @@
+"""
+ChatBot - Configuration Settings for the TinyLLM Chatbot.
+
+It includes settings for the OpenAI API, LiteLLM Proxy, Weaviate, and other options.
+It also includes settings for the chatbot's behavior, such as temperature, max tokens, and more.
+"""
+
+VERSION = "v0.16.0"
+
+# Imports
+import os
+import json
+import time
+import logging
+
+# Set up logging
+logging.basicConfig(level=logging.INFO, 
+                    format='%(asctime)s %(levelname)s %(message)s', 
+                    datefmt='%Y-%m-%d %H:%M:%S')
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
+logger.info("TinyLLM %s" % VERSION)
+
+def log(text):
+    logger.info(text)
+
+def debug(text):
+    logger.debug(text)
+
+# Configuration Settings
+API_KEY = os.environ.get("OPENAI_API_KEY", "Asimov-3-Laws")                 # Required, use bogus string for local LLMs
+API_BASE = os.environ.get("OPENAI_API_BASE", "http://localhost:8000/v1")    # Required, use https://api.openai.com for OpenAI
+LITELLM_PROXY = os.environ.get("LITELLM_PROXY", None)                       # Optional - LITELLM Proxy URL
+LITELLM_KEY = os.environ.get("LITELLM_KEY", "")                             # Optional - LITELLM Secret Key - Begins with sk-
+AGENTNAME = os.environ.get("AGENT_NAME", "")                                # Set the name of your bot
+MYMODEL = os.environ.get("LLM_MODEL", "models/7B/gguf-model.bin")           # Pick model to use e.g. gpt-3.5-turbo for OpenAI
+DEBUG = os.environ.get("DEBUG", "false").lower() == "true"                  # Set to True to enable debug mode
+MAXCLIENTS = int(os.environ.get("MAXCLIENTS", 1000))                        # Maximum number of concurrent clients
+MAXTOKENS = int(os.environ.get("MAXTOKENS", 0))                             # Maximum number of tokens to send to LLM for RAG
+TEMPERATURE = float(os.environ.get("TEMPERATURE", 0.0))                     # LLM temperature
+PORT = int(os.environ.get("PORT", 5000))                                    # Port to listen on
+PROMPT_FILE = os.environ.get("PROMPT_FILE", f".tinyllm/prompts.json")       # File to store system prompts
+PROMPT_RO = os.environ.get("PROMPT_RO", "false").lower() == "true"          # Set to True to enable read-only prompts
+USE_SYSTEM = os.environ.get("USE_SYSTEM", "false").lower == "true"          # Use system in chat prompt if True
+TOKEN = os.environ.get("TOKEN", "secret")                                   # Secret TinyLLM token for admin functions
+ONESHOT = os.environ.get("ONESHOT", "false").lower() == "true"              # Set to True to enable one-shot mode
+RAG_ONLY = os.environ.get("RAG_ONLY", "false").lower() == "true"            # Set to True to enable RAG only mode
+EXTRA_BODY = os.environ.get("EXTRA_BODY", None)                             # Extra body parameters for OpenAI API
+TOXIC_THRESHOLD = float(os.environ.get("TOXIC_THRESHOLD", 99))              # Toxicity threshold for responses 0-1 or 99 disable
+THINKING = os.environ.get("THINKING", "false").lower() == "true"            # Set to True to enable thinking mode by default
+THINK_FILTER = os.environ.get("THINK_FILTER", "false").lower() == "true"    # Set to True to enable thinking filter
+MAX_IMAGES = int(os.environ.get("MAX_IMAGES", 1))                           # Maximum number of images to keep in context
+SEARXNG = os.environ.get("SEARXNG", "http://localhost:8080")                # SearXNG URL for internet search engine  (Optional)
+INTENT_ROUTER = os.environ.get("INTENT_ROUTER", "false").lower() == "true"  # Set to True to enable intent detection
+WEB_SEARCH = os.environ.get("WEB_SEARCH", "false").lower() == "true"        # Set to True to enable web search for all queries
+
+# Convert EXTRA_BODY to dictionary if it is proper JSON
+if EXTRA_BODY:
+    try:
+        EXTRA_BODY = json.loads(EXTRA_BODY)
+    except:
+        log("EXTRA_BODY is not valid JSON")
+        EXTRA_BODY = {}
+else:
+    if API_BASE.startswith("https://api.openai.com") or LITELLM_PROXY:
+        EXTRA_BODY = {}
+    else:
+        # Extra stop tokens are needed for some non-OpenAI LLMs
+        EXTRA_BODY = {"stop_token_ids":[128001, 128009]}
+
+# LiteLLM Proxy
+if LITELLM_PROXY:
+    log(f"Using LiteLLM Proxy at {LITELLM_PROXY}")
+    API_BASE = LITELLM_PROXY
+    API_KEY = LITELLM_KEY
+
+# RAG Configuration Settings
+WEAVIATE_HOST = os.environ.get("WEAVIATE_HOST", "")                         # Empty = no Weaviate support
+WEAVIATE_GRPC_HOST = os.environ.get("WEAVIATE_GRPC_HOST", WEAVIATE_HOST)    # Empty = no Weaviate gRPC support
+WEAVIATE_PORT = os.getenv('WEAVIATE_PORT', '8080')
+WEAVIATE_GRPC_PORT = os.getenv('WEAVIATE_GRPC_PORT', '50051')
+WEAVIATE_LIBRARY = os.environ.get("WEAVIATE_LIBRARY", "tinyllm")            # Weaviate library to use
+WEAVIATE_AUTH_KEY = os.getenv('WEAVIATE_AUTH_KEY', None)                    # Weaviate Auth Key
+RESULTS = int(os.environ.get("RESULTS", 1))                                 # Number of results to return from RAG query
+ALPHA_KEY = os.environ.get("ALPHA_KEY", "alpha_key")                        # Optional - Alpha Vantage API Key
+UPLOAD_FOLDER = os.environ.get("UPLOAD_FOLDER", "/tmp")                     # Folder to store uploaded documents
+
+# Debug Mode
+if DEBUG:
+    logger.setLevel(logging.DEBUG)
+    log("Debug mode enabled.")
+    # Display all default settings
+    debug("Configuration Settings:")
+    vars = globals()
+    for n in list(vars):
+        if n.isupper(): 
+            if vars[n] and n in ["API_KEY", "TOKEN", "WEAVIATE_AUTH_KEY", "ALPHA_KEY"]:
+                debug(f"   {n}: {'*' * len(vars[n])}")
+            else:
+                debug(f"   {n}: {vars[n]}")
+
+# Log ONE_SHOT mode
+if ONESHOT:
+    log("ONESHOT mode enabled.")
+
+# Globals
+client = {}
+prompts = {}
+stats = {
+    "start_time": time.time(),
+    "errors": 0,
+    "ask": 0,
+    "ask_llm": 0,
+    "ask_context": 0,
+}
+llm_stream = None
