--- chatbot/version.py
@@ -1,1 +1,0 @@
-VERSION = "v0.15.25"
